+++ python/sierrapy/commands/__init__.py
@@ -3,5 +3,6 @@ from . import fasta  # noqa
 from . import mutations  # noqa
 from . import patterns  # noqa
 from . import introspection  # noqa
+from . import recipe  # noqa
 
 __all__ = ['cli']