+++ python/sierrapy/recipes/__init__.py
@@ -1,0 +1,3 @@
+from .alignment import alignment
+
+__all__ = ['alignment']